backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
pymongo==4.4.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
geopy==2.4.0
python-dotenv==1.0.0